Add useQuote hook.
