+++ src/components/Quotebox.jsx
@@ -1,39 +1,11 @@
 import React from 'react';
-import { useState, useEffect } from 'react';
 import SocialMenu from './SocialMenu';
 import { PHASES } from '../App';
-
-const QUOTESURL = "https://api.quotable.io/random?tags=";
+import { useQuote } from '../hooks/useQuote.js'
 
 export default function Quotebox({ tags, setPhase }) {
-  const [quote, setQuote] = useState({});
-
-  //Fetch quote
-  async function fetchQuote(){
-    const FILTERS = tags.filter((tag) => tag.active === true).map((tag) => tag.slug).join('|');
-    console.log(`${QUOTESURL}${FILTERS}`)
-    try {
-      const response = await fetch(`${QUOTESURL}${FILTERS}`);
-      if (!response.ok) {
-        throw new Error(`Response status: ${response.status}`);
-      }
+  const [quote, fetchQuote] = useQuote(tags);
   
-      const result = await response.json();
-      setQuote({
-        id: result._id,
-        content: result.content,
-        author: result.author,
-        length: result.length
-      });
-    } catch (error) {
-      console.error(error.message);
-    }
-  }
-
-  useEffect(() => {
-    fetchQuote();
-  }, []);
-
   return (
     <div className="box">
       <div className="quote">
@@ -49,7 +21,7 @@ export default function Quotebox({ tags, setPhase }) {
         </div>
         <div className='settings'>
           <SocialMenu quote={quote}/>
-          <button onClick={() => fetchQuote()}>New Quote</button>
+          <button onClick={() => fetchQuote(tags)}>New Quote</button>
         </div>
       </div>
     </div>
